fix(reliability): extract route handlers into _handler.ts files
Next.js App Router rejects any export from a `route.ts` file that
isn't on its allow-list (GET/POST/etc. + `runtime`/`dynamic`/etc.),
so the previous `__handleForTests` re-export tripped the type-gen
check during `next build`:

  Type error: Route ".../probe/route.ts" does not match the required
  types of a Next.js Route. "__handleForTests" is not a valid Route
  export field.

Move `handleProbe` / `handleSnapshots` and the `__handleForTests`
test shim into sibling `_handler.ts` files (underscore prefix is
the same convention already used for `_auth.ts`). The route files
shrink to just the allowed Next.js exports + a thin delegation to
the handler.

Update both route tests to import `__handleForTests` from
`_handler` instead of `route`.

Validation
- pnpm --filter @vitalcv/web exec vitest run __tests__/source-health/
  -> 88/88 pass (9 files)
- pnpm turbo run build --filter @vitalcv/web --force
  -> 13/13 tasks (full Next.js build with TypeScript + ESLint
  enforcement on)
